#956 Phase 7: extract cos/queue_service.rs from tx.rs
Per docs/pr/956-phase7-queue-service/plan.md (v3, post Codex 2 rounds
+ Gemini 2 rounds). Largest move of the campaign — ~2261 LOC covering
the full per-byte hot-path chain.

Move list (7 types + 38 fns):
- Types: CoSBatch, CoSServicePhase, ExactCoSQueueKind,
  ExactCoSQueueSelection, ExactCoSScratchBuild, DrainedQueueRef,
  ParkReason
- Fns: drain_shaped_tx, cos_batch_queue_ref, build_nonexact_cos_batch,
  service_exact_guarantee_queue_direct(_with_info), 5 select_*_batch
  variants, 4 service_exact_*_queue_direct(_flow_fair) variants,
  4 drain_exact_*_to_scratch, 4 settle_*, 2 release_*_scratch,
  2 restore_*_scratch_to_queue_head_flow_fair,
  subtract_direct_cos_queue_bytes, build_cos_batch_from_queue,
  submit_cos_batch, cos_batch_tx_made_progress, 2 quantum helpers,
  estimate_cos_queue_wakeup_tick, count_park_reason, park_cos_queue,
  assign_local_dscp_rewrite, assign_prepared_dscp_rewrite,
  restore_cos_local_items, restore_cos_prepared_items.

#[inline] added on per-byte/per-batch hot path: drain_shaped_tx,
all selectors, all service_*_direct, build_cos_batch_from_queue,
build_nonexact_cos_batch, cos_batch_queue_ref, submit_cos_batch,
cos_batch_tx_made_progress, cos_*_quantum_bytes, subtract_*,
assign_*_dscp. Existing #[inline] on count_park_reason preserved.
Larger bodies (drain_*_to_scratch, settle_*) skip #[inline].

Visibility:
- pub(in crate::afxdp): drain_shaped_tx + 6 enum/struct types +
  9 cfg-gated test-touched items + 2 quantum helpers + park_cos_queue
- File-private: dispatch internals (build_*, submit_cos_batch,
  cos_batch_tx_made_progress, settle/release/restore helpers)

Back-edges to tx.rs (deferred TX-completion + worker-binding helpers):
apply_cos_*_result (3), prime_cos_root_for_service,
advance_cos_timer_wheel, restore_cos_*_inner (2), transmit_batch,
transmit_prepared_queue, reap_tx_completions, refresh_cos_interface_activity,
cos_queue_dscp_rewrite, count_tx_ring_full_submit_stall,
recycle_cancelled_prepared_offset, remember_prepared_recycle,
maybe_wake_tx, COS_GUARANTEE_VISIT_NS, COS_GUARANTEE_QUANTUM_*,
COS_SURPLUS_ROUND_QUANTUM_BYTES, TxError. Visibility on all bumped
to pub(in crate::afxdp). frame::frame_has_tcp_rst,
frame::apply_dscp_rewrite_to_frame, neighbor::monotonic_nanos,
tx::stamp_submits also bumped.

Net change: -2165 lines from tx.rs (16122 -> 13957).

Tests: cargo test --bins -> 865 passed, 0 failed, 2 ignored.

Reviewers signed off:
- Codex rounds 1-2 caught scope incompleteness (enums, scratch +
  settle helpers, submit + dscp helpers); v3 expanded.
- Gemini rounds 1-2: PLAN-READY at v2 (accepting hot-path
  inline + cohesion arguments).
